Move dialogue to module within src/
